Avoid using hard coded network port within UT; (#7526)
If we use hard coded network port within UT, there would be conflict under some circumstances.

Approved by: @XuPeng-SH, @zhangxu19830126, @fengttt, @sukki37
